Read files with short map ID 'MAP' without errors or warnings
These files are valid according to the MRC2014 paper but not the
MRC2014 spec on the website. We should read them correctly without
needing to use permissive mode. However they should still fail in
validation, where we should be as strict as possible to help
developers to write files that are fully valid according to all
versions of the specification.