Allow failures for ARM builds
I can't get tox-travis to exclude the numpy-1.18 environments on arm64
which don't build correctly. This seems like a reasonable fix for now,
and the osx build has been quite reliable so it should be ok to stop
allowing failures for that. (It's not possible to allow failures in two
different categories.)