Use explicit endianness for FEI extended header dtypes
This fixes https://github.com/ccpem/mrcfile/issues/35

The FEI documentation is unclear about endianness. It's quite likely
that no MRC files will ever exist that have big-endian data in the FEI
extended header so perhaps we could have just assumed little-endianness
in all cases, but nonetheless it felt safer to make the extended
header's endianness match the standard header (except for the bitmask
fields that are specified as little-endian and required a bit of
complexity to get right).