Update statistics handling
Using float64 or complex128 accumulators for the RMS slows down calculations
a lot. Instead we just use float32 or complex64, accepting that the accuracy
will be slightly lower but that shouldn't be a problem since these summary
fields are not very important. To allow for that, we add a bit (1%) of
tolerance when validating the mean and RMS values.

Also, we now raise an error an all unexpected warnings in the test suite. One
of the test data arrays included some very large float32 values that triggered
overflow in the std calculation, but that is expected to happen very rarely in
real files so this commit reduces the magnitude of values in the test arrays
so std does not overflow.

If a user does now calculate the header stats for a file which causes overflow,
they will see the RuntimeWarning from numpy and the header field will end up
set to 'inf'.

https://github.com/ccpem/mrcfile/issues/45