Raise errors if arrays are too large
If the size of a new extended header or data array has a dimension
too large to be stored in the header (where the dimensions are stored
as int32 values) a ValueError is now raised.

Previously, the excessive value would overflow leading to a negative
size being stored in the header.